Relocate common classes to MuonGEMHits. Remove CSC legacy codes.
