+++ Validation/MuonGEMDigis/interface/GEMCSCPadDigiValidation.h
@@ -10,7 +10,7 @@
 #include "FWCore/Framework/interface/EDAnalyzer.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 
-#include "Validation/MuonGEMDigis/interface/GEMBaseValidation.h"
+#include "Validation/MuonGEMHits/interface/GEMBaseValidation.h"
 
 
 class GEMCSCPadDigiValidation : public GEMBaseValidation
