+++ Validation/MuonGEMDigis/interface/GEMTrackMatch.h
@@ -21,7 +21,7 @@
 #include "Geometry/CommonTopologies/interface/StripTopology.h"
 
 
-#include "Validation/MuonGEMDigis/interface/SimTrackDigiMatchManager.h"
+#include "Validation/MuonGEMHits/interface/SimTrackMatchManager.h"
 
 
 class GEMTrackMatch
