--- Validation/MuonGEMDigis/src/SimTrackDigiMatchManager.cc
@@ -1,9 +1,0 @@
-#include "Validation/MuonGEMDigis/interface/SimTrackDigiMatchManager.h"
-
-SimTrackDigiMatchManager::SimTrackDigiMatchManager(const SimTrack& t, const SimVertex& v,
-      const edm::ParameterSet& ps, const edm::Event& ev, const edm::EventSetup& es)
-: simhits_(t, v, ps, ev, es)
-, gem_digis_(simhits_)
-{}
-
-SimTrackDigiMatchManager::~SimTrackDigiMatchManager() {}
