+++ Validation/MuonGEMHits/interface/BaseMatcher.h
@@ -28,10 +28,6 @@ class BaseMatcher
 {
 public:
   
-  /// CSC chamber types, according to CSCDetId::iChamberType()
-  enum CSCType {CSC_ALL = 0, CSC_ME1a, CSC_ME1b, CSC_ME12, CSC_ME13,
-      CSC_ME21, CSC_ME22, CSC_ME31, CSC_ME32, CSC_ME41, CSC_ME42};
-
 
   BaseMatcher(const SimTrack& t, const SimVertex& v,
       const edm::ParameterSet& ps, const edm::Event& ev, const edm::EventSetup& es);
@@ -51,9 +47,6 @@ public:
   const edm::Event& event() const {return ev_;}
   const edm::EventSetup& eventSetup() const {return es_;}
 
-  /// check if CSC chamber type is in the used list
-  bool useCSCChamberType(int csc_type);
-  
   void setVerbose(int v) { verbose_ = v; }
   int verbose() const { return verbose_; }
 
@@ -78,8 +71,6 @@ private:
 
   int verbose_;
 
-  // list of CSC chamber types to use
-  bool useCSCChamberTypes_[11];
 
   edm::ESHandle<MagneticField> magfield_;
   edm::ESHandle<Propagator> propagator_;
