+++ Validation/MuonGEMHits/interface/GEMBaseValidation.h
@@ -10,19 +10,13 @@
 #include "FWCore/Framework/interface/EDAnalyzer.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 
-#include "SimDataFormats/TrackingHit/interface/PSimHitContainer.h"
-#include "SimDataFormats/Track/interface/SimTrackContainer.h"
-
 class GEMBaseValidation
 {
 public:
   GEMBaseValidation(DQMStore* dbe,
                          const edm::InputTag & inputTag);
   virtual ~GEMBaseValidation();
-  //virtual void analyze(const edm::Event& e, const edm::EventSetup&) = 0;
-
   void setGeometry(const GEMGeometry* geom) { theGEMGeometry = geom; }
-  // void set SimHitMap(const PSimHitMap* simHitMap) { theSimHitMap = simHitMap; } 
 
 
  protected:
