+++ Validation/MuonGEMHits/interface/SimTrackMatchManager.h
@@ -12,8 +12,9 @@
 
 */
 
-#include "Validation/MuonGEMDigis/interface/BaseMatcher.h"
-#include "Validation/MuonGEMDigis/interface/SimHitMatcher.h"
+#include "Validation/MuonGEMHits/interface/BaseMatcher.h"
+#include "Validation/MuonGEMHits/interface/SimHitMatcher.h"
+#include "Validation/MuonGEMHits/interface/GEMDigiMatcher.h"
 
 class SimTrackMatchManager
 {
@@ -25,10 +26,12 @@ public:
   ~SimTrackMatchManager();
 
   const SimHitMatcher& simhits() const {return simhits_;}
+  const GEMDigiMatcher& gemDigis() const {return gem_digis_;}
   
 private:
 
   SimHitMatcher simhits_;
+  GEMDigiMatcher gem_digis_;
 };
 
 #endif
