+++ Validation/MuonGEMHits/src/BaseMatcher.cc
@@ -1,4 +1,4 @@
-#include "Validation/MuonGEMDigis/interface/BaseMatcher.h"
+#include "Validation/MuonGEMHits/interface/BaseMatcher.h"
 
 #include "TrackingTools/TrajectoryState/interface/TrajectoryStateOnSurface.h"
 #include "TrackingTools/Records/interface/TrackingComponentsRecord.h"
@@ -10,15 +10,6 @@ BaseMatcher::BaseMatcher(const SimTrack& t, const SimVertex& v,
       const edm::ParameterSet& ps, const edm::Event& ev, const edm::EventSetup& es)
 : trk_(t), vtx_(v), conf_(ps), ev_(ev), es_(es), verbose_(0)
 {
-  // list of CSC chamber type numbers to use
-  std::vector<int> csc_types = conf().getUntrackedParameter<std::vector<int> >("useCSCChamberTypes", std::vector<int>() );
-  for (int i=0; i <= CSC_ME42; ++i) useCSCChamberTypes_[i] = false;
-  for (auto t: csc_types)
-  {
-    if (t >= 0 && t <= CSC_ME42) useCSCChamberTypes_[t] = 1;
-  }
-  // empty list means use all the chamber types
-  if (csc_types.empty()) useCSCChamberTypes_[CSC_ALL] = 1;
 
   // Get the magnetic field
   es.get< IdealMagneticFieldRecord >().get(magfield_);
@@ -34,11 +25,6 @@ BaseMatcher::~BaseMatcher()
 }
 
 
-bool BaseMatcher::useCSCChamberType(int csc_type)
-{
-  if (csc_type < 0 || csc_type > CSC_ME42) return false;
-  return useCSCChamberTypes_[csc_type];
-}
 
 
 GlobalPoint
